+++ src/idvc/dvc_interface.py
@@ -277,9 +277,10 @@ class MainWindow(QMainWindow):
 
         self.viewer3D_dock = QDockWidget("3D View")
         self.viewer3D_dock.setObjectName("3DImageView")
+        self.viewer3D_dock.setMinimumHeight(300)
         self.viewer3D_dock.setWidget(self.vis_widget_3D)
         self.viewer3D_dock.setAllowedAreas(Qt.LeftDockWidgetArea)
-        self.viewer3D_dock.setFeatures(QDockWidget.NoDockWidgetFeatures)   
+        self.viewer3D_dock.setFeatures(QDockWidget.DockWidgetFloatable)   
 
         #Tabifies dockwidgets in LeftDockWidgetArea:
         prev = None
@@ -499,7 +500,7 @@ class MainWindow(QMainWindow):
 
         self.help_label = QLabel(groupBox)
         scroll_area_widget = QScrollArea()
-        
+    
         self.help_label.setWordWrap(True)
         self.help_label.setText(self.help_text[0])
 
@@ -2218,6 +2219,7 @@ It is used as a global starting point and a translation reference."
         widgetno += 1
 
         mp_widgets['mask_downsampled_coords_warning'] = QLabel(groupBox)
+        mp_widgets['mask_downsampled_coords_warning'].setWordWrap(True)
         mp_widgets['mask_downsampled_coords_warning'].setText("Note: if your image has been downsampled, the number of slices is in the coordinates of the downsampled image.")
         formLayout.setWidget(widgetno, QFormLayout.FieldRole, mp_widgets['mask_downsampled_coords_warning'])
         widgetno += 1
@@ -2644,7 +2646,6 @@ It is used as a global starting point and a translation reference."
 # Point Cloud Panel:
 
     def CreatePointCloudPanel(self):
-
         self.pointCloudDockWidget = QDockWidget(self)
         self.pointCloudDockWidget.setFeatures(QDockWidget.NoDockWidgetFeatures)
         self.pointCloudDockWidget.setWindowTitle('4 - Point Cloud')
@@ -2655,11 +2656,18 @@ It is used as a global starting point and a translation reference."
 
 
         # Add vertical layout to dock contents
+
         self.graphDockVL = QVBoxLayout(self.pointCloudDockWidgetContents)
         self.graphDockVL.setContentsMargins(0, 0, 0, 0)
 
+        
+
         # Create widget for dock contents
         self.dockWidget = QWidget(self.pointCloudDockWidgetContents)
+
+        scroll_area_point_cloud = QScrollArea()
+        scroll_area_point_cloud.setWidgetResizable(True)
+        scroll_area_point_cloud.setWidget(self.dockWidget)
 
         # Add vertical layout to dock widget
         self.graphWidgetVL = QVBoxLayout(self.dockWidget)
@@ -2929,7 +2937,7 @@ A 3D pointcloud is created within the full extent of the mask."
         widgetno += 1
         # Add elements to layout
         self.graphWidgetVL.addWidget(self.graphParamsGroupBox)
-        self.graphDockVL.addWidget(self.dockWidget)
+        self.graphDockVL.addWidget(scroll_area_point_cloud)
         self.pointCloudDockWidget.setWidget(self.pointCloudDockWidgetContents)
         self.addDockWidget(QtCore.Qt.LeftDockWidgetArea, self.pointCloudDockWidget)
         widgetno += 1